requirements.txt
pyspark==3.5.1
pymongo==4.6.1
mysql-connector-python==8.0.33
python-dotenv==1.0.1
pandas==2.2.2
prophet==1.1.5
black==24.4.2
